+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.4.1 (unreleased)
 
+- Aliased `apply` to `map` for `Series`
 - Improved `any` and `all` for `Series`
 
 ## 0.4.0 (2023-04-01)